+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Support for Symfony 4.x
 ### Removed
 - Support for Symfony 3.x
+### Fixed
+- Fixed custom sorting issue in Tree Admin caused by sonata-project/doctrine-orm-admin-bundle 3.15
 
 ## 5.2.0 - 2020-05-14
 ### Fixed